+++ web/share-cart/app/root.tsx
@@ -9,7 +9,7 @@ import "./tailwind.css";
 
 export function Layout({ children }: { children: React.ReactNode }) {
 	return (
-		<html lang="en">
+		<html lang="ja">
 			<head>
 				<meta charSet="utf-8" />
 				<meta name="viewport" content="width=device-width, initial-scale=1" />
