--- web/share-cart/app/routes/auth.auth0.callback.tsx
@@ -1,10 +1,0 @@
-import type { LoaderFunctionArgs } from "@remix-run/node";
-
-import { authenticator } from "~/service/auth.server";
-
-export const loader = ({ request }: LoaderFunctionArgs) => {
-	return authenticator.authenticate("auth0", request, {
-		successRedirect: "/auth/user/register",
-		failureRedirect: "/auth/login",
-	});
-};
