--- web/share-cart/app/routes/auth.login.tsx
@@ -1,9 +1,0 @@
-import { Form } from "@remix-run/react";
-
-export default function Login() {
-	return (
-		<Form action="/auth/auth0/login" method="post">
-			<button type="submit">Login with Auth0</button>
-		</Form>
-	);
-}
