--- web/share-cart/app/routes/auth.logout.tsx
@@ -1,9 +1,0 @@
-import { Form } from "@remix-run/react";
-
-export default function Logout() {
-	return (
-		<Form action="/auth/auth0/logout" method="post">
-			<button type="submit">Logout with Auth0</button>
-		</Form>
-	);
-}
