+++ web/share-cart/package.json
@@ -20,7 +20,8 @@
 		"react": "^18.2.0",
 		"react-dom": "^18.2.0",
 		"remix-auth": "^3.7.0",
-		"remix-auth-auth0": "^1.10.0"
+		"remix-auth-auth0": "^1.10.0",
+		"remix-routes": "^1.7.6"
 	},
 	"devDependencies": {
 		"@biomejs/biome": "^1.8.3",
@@ -30,6 +31,7 @@
 		"autoprefixer": "^10.4.20",
 		"dependency-cruiser": "^16.3.10",
 		"postcss": "^8.4.38",
+		"remix-flat-routes": "^0.6.5",
 		"tailwindcss": "^3.4.4",
 		"typescript": "^5.1.6",
 		"vite": "^5.1.0",
