+++ web/share-cart/vite.config.ts
@@ -1,4 +1,6 @@
 import { vitePlugin as remix } from "@remix-run/dev";
+import { flatRoutes } from 'remix-flat-routes';
+import { remixRoutes } from 'remix-routes/vite';
 import { defineConfig } from "vite";
 import tsconfigPaths from "vite-tsconfig-paths";
 
@@ -10,7 +12,12 @@ export default defineConfig({
 				v3_relativeSplatPath: true,
 				v3_throwAbortReason: true,
 			},
+			ignoredRouteFiles: ['**/*'],
+			routes: async defineRoutes => {
+				return flatRoutes('routes', defineRoutes)
+			},
 		}),
 		tsconfigPaths(),
+		remixRoutes(),
 	],
 });
